Moved Neural Network Functions All Neural Network Functions have been split into Fitness,Mutate, and NeuralNetwork
