--- Lua/NES/temp.pool
@@ -1,608 +1,0 @@
-0
-0
-25
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.315789
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.475
-disable
-0.38
-done
-3
-170 1000003 -1.3364055299539 9 1
-91 1000004 -1.9400616473891 8 1
-132 1000004 1.4357737968078 7 0
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.285
-link
-2.10526
-enable
-0.210526
-bias
-0.421052
-node
-0.526315
-disable
-0.38
-done
-2
-62 1000003 0.43073824274422 10 1
-142 1000005 -0.79219946897794 11 1
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.285
-link
-1.9
-enable
-0.210526
-bias
-0.38
-node
-0.526315
-disable
-0.421052
-done
-3
-108 1000001 -0.93514816736351 12 1
-148 1000003 -0.49653614917447 13 0
-170 1000003 -1.9648426770837 14 1
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.315789
-link
-2.10526
-enable
-0.19
-bias
-0.38
-node
-0.526315
-disable
-0.38
-done
-5
-132 1000003 0.3961912900174 15 0
-130 1000005 0.28962065492721 16 1
-170 1000001 1.2115237891781 17 0
-170 171 1 18 1
-171 1000001 1.2115237891781 19 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.285
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.475
-disable
-0.421052
-done
-2
-122 1000004 -0.29377117221595 20 0
-1000003 1000002 -1.3870662556841 21 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.285
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.475
-disable
-0.38
-done
-2
-134 1000002 -0.098818933683279 22 1
-65 1000006 -1.8596148564104 23 0
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.315789
-link
-1.9
-enable
-0.19
-bias
-0.38
-node
-0.526315
-disable
-0.38
-done
-4
-1000005 1000004 0.95675527207251 24 1
-35 1000002 1.3577684865871 25 1
-35 171 1 26 1
-171 1000002 1.3577684865871 27 1
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.315789
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.526315
-disable
-0.38
-done
-2
-20 1000002 -0.74373607593005 28 1
-170 1000003 0.93234046449171 29 1
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.285
-link
-1.9
-enable
-0.19
-bias
-0.38
-node
-0.526315
-disable
-0.421052
-done
-5
-98 1000003 -1.5449079866939 30 1
-133 1000002 0.17377239295633 31 0
-170 1000004 0.7848750267037 32 1
-133 171 1 33 1
-171 1000002 0.17377239295633 34 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.315789
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.526315
-disable
-0.421052
-done
-2
-32 1000006 -0.17023224585711 35 1
-170 1000003 -0.24909207434309 36 1
-0
-0
-1
-0
-171
-connections
-0.665
-step
-0.285
-link
-2.10526
-enable
-0.19
-bias
-0.38
-node
-0.526315
-disable
-0.421052
-done
-5
-106 1000005 1.4191717276528 37 0
-110 1000002 0.26300851466414 38 1
-170 1000005 1.6203497421186 39 1
-106 171 1 40 1
-171 1000005 1.4191717276528 41 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.285
-link
-1.9
-enable
-0.19
-bias
-0.421052
-node
-0.526315
-disable
-0.421052
-done
-1
-27 1000005 1.4709311197241 42 1
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.285
-link
-2.10526
-enable
-0.19
-bias
-0.38
-node
-0.475
-disable
-0.421052
-done
-4
-28 1000004 -1.6826075014496 43 0
-113 1000002 0.18036439100314 44 0
-28 171 1 45 1
-171 1000004 -1.6826075014496 46 1
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.285
-link
-2.10526
-enable
-0.210526
-bias
-0.421052
-node
-0.526315
-disable
-0.38
-done
-5
-12 1000002 0.098330637531663 47 1
-169 1000003 1.5025482955412 48 1
-170 1000005 -1.0916470839564 49 0
-170 171 1 50 1
-171 1000005 -1.0916470839564 51 1
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.285
-link
-1.9
-enable
-0.19
-bias
-0.421052
-node
-0.526315
-disable
-0.38
-done
-2
-107 1000001 -1.9764397106845 52 1
-18 1000005 1.4529862361522 53 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.315789
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.475
-disable
-0.38
-done
-2
-168 1000005 -1.8886684774316 54 1
-10 1000005 -1.4734946745201 55 1
-0
-0
-1
-0
-170
-connections
-0.736841
-step
-0.285
-link
-1.9
-enable
-0.210526
-bias
-0.421052
-node
-0.475
-disable
-0.38
-done
-2
-61 1000004 -0.52974028748436 56 1
-142 1000002 0.10614337595752 57 0
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.285
-link
-1.9
-enable
-0.210526
-bias
-0.38
-node
-0.475
-disable
-0.421052
-done
-2
-40 1000004 -1.557481612598 58 1
-24 1000004 1.1524399548326 59 1
-0
-0
-1
-0
-171
-connections
-0.665
-step
-0.315789
-link
-1.9
-enable
-0.210526
-bias
-0.38
-node
-0.475
-disable
-0.421052
-done
-5
-15 1000005 -0.29279457991272 60 0
-46 1000006 -1.0689413129063 61 1
-170 1000006 -1.679555650502 62 1
-15 171 1 63 1
-171 1000005 -0.29279457991272 64 1
-0
-0
-1
-0
-171
-connections
-0.665
-step
-0.285
-link
-2.10526
-enable
-0.210526
-bias
-0.421052
-node
-0.526315
-disable
-0.421052
-done
-4
-1 1000005 1.8192083498642 65 1
-43 1000005 -0.97946104312265 66 1
-43 171 1 67 1
-171 1000005 -0.97946104312265 68 1
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.315789
-link
-1.9
-enable
-0.19
-bias
-0.38
-node
-0.475
-disable
-0.38
-done
-4
-151 1000003 1.2442396313364 69 1
-122 1000002 0.12054811243019 70 0
-122 171 1 71 1
-171 1000002 0.12054811243019 72 1
-0
-0
-1
-0
-171
-connections
-0.665
-step
-0.315789
-link
-2.10526
-enable
-0.210526
-bias
-0.38
-node
-0.526315
-disable
-0.421052
-done
-5
-26 1000006 0.47993408001953 73 1
-163 1000001 0.77816095461898 74 1
-170 1000005 1.6091189306314 75 0
-170 171 1 76 1
-171 1000005 1.6091189306314 77 0
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.315789
-link
-2.10526
-enable
-0.210526
-bias
-0.38
-node
-0.475
-disable
-0.38
-done
-2
-72 1000004 1.9704580828272 78 1
-45 1000004 -1.0092471083712 79 0
-0
-0
-1
-0
-171
-connections
-0.736841
-step
-0.315789
-link
-2.10526
-enable
-0.19
-bias
-0.421052
-node
-0.526315
-disable
-0.38
-done
-5
-75 1000006 0.16229743339335 80 1
-95 1000004 -0.41218298898282 81 1
-170 1000004 -1.7188634907071 82 1
-95 171 1 83 0
-171 1000004 -0.41218298898282 84 1
-0
-0
-1
-0
-170
-connections
-0.665
-step
-0.315789
-link
-2.10526
-enable
-0.19
-bias
-0.421052
-node
-0.475
-disable
-0.421052
-done
-2
-168 1000006 -0.4331797235023 85 0
-33 1000006 0.29779961546678 86 1
